+++ test/grouping.jl
@@ -1708,8 +1708,8 @@ end
 @testset "Allow returning DataFrame() or NamedTuple() to drop group" begin
     N = 4
     for (i, x1) in enumerate(collect.(Iterators.product(repeat([[true, false]], N)...))),
-        er in (DataFrame(), view(DataFrame(ones(2,2)), 2:1, 2:1),
-               view(DataFrame(ones(2,2)), 1:2, 2:1),
+        er in (DataFrame(), view(DataFrame(Tables.table(ones(2,2), header=[:x1, :x2])), 2:1, 2:1),
+               view(DataFrame(Tables.table(ones(2,2), header=[:x1, :x2])), 1:2, 2:1),
                NamedTuple(), rand(0,0), rand(5,0),
                DataFrame(x1=Int[]), DataFrame(x1=Any[]),
                (x1=Int[],), (x1=Any[],), rand(0,1)),
@@ -1875,7 +1875,7 @@ end
               DataFrame(g=1:100, g_function=1:100)
     end
 
-    df_ref = DataFrame(rand(10, 4))
+    df_ref = DataFrame(Tables.table(rand(10, 4), header=[:x1, :x2, :x3, :x4]))
     df_ref.g = shuffle!([1,2,2,3,3,3,4,4,4,4])
 
     for i in 0:nrow(df_ref), dosort in [true, false], dokeepkeys in [true, false]
@@ -1892,7 +1892,7 @@ end
 end
 
 @testset "passing columns" begin
-    df = DataFrame(rand(10, 4))
+    df = DataFrame(Tables.table(rand(10, 4), header=[:x1, :x2, :x3, :x4]))
     df.g = shuffle!([1,2,2,3,3,3,4,4,4,4])
     gdf = groupby_checked(df, :g)
 
@@ -2817,9 +2817,10 @@ end
                 maximum, minimum, maximum∘skipmissing, minimum∘skipmissing,
                 first, last, length, first∘skipmissing, last∘skipmissing),
         col in ([ones(2,2), zeros(2,2), ones(2,2)], [ones(2,2), zeros(2,2), missing],
-                [DataFrame(ones(2,2)), DataFrame(zeros(2,2)), DataFrame(ones(2,2))],
-                [DataFrame(ones(2,2)), DataFrame(zeros(2,2)), ones(2,2)],
-                [DataFrame(ones(2,2)), DataFrame(zeros(2,2)), missing],
+                [DataFrame(Tables.table(ones(2,2))), DataFrame(Tables.table(zeros(2,2))),
+                DataFrame(Tables.table(ones(2,2)))], [DataFrame(Tables.table(ones(2,2))),
+                DataFrame(Tables.table(zeros(2,2))), ones(2,2)],
+                [DataFrame(Tables.table(ones(2,2))), DataFrame(Tables.table(zeros(2,2))), missing],
                 [(a=1, b=2), (a=3, b=4), (a=5, b=6)], [(a=1, b=2), (a=3, b=4), missing])
         gdf = groupby_checked(DataFrame(g=[1, 1, 1], x=col), :g)
         if fun === length