+++ test/iteration.jl
@@ -94,7 +94,7 @@ end
 end
 
 @testset "SubDataFrame" begin
-    df = DataFrame([11:16 21:26 31:36 41:46])
+    df = DataFrame(Tables.table([11:16 21:26 31:36 41:46]))
     sdf = view(df, [3,1,4], [3,1,4])
     @test sdf == df[[3,1,4], [3,1,4]]
     @test eachrow(sdf) == eachrow(df[[3,1,4], [3,1,4]])
@@ -104,7 +104,7 @@ end
 end
 
 @testset "parent mutation" begin
-    df = DataFrame([11:16 21:26 31:36 41:46])
+    df = DataFrame(Tables.table([11:16 21:26 31:36 41:46]))
     sdf = view(df, [3,1,4], [3,1,4])
     erd = eachrow(df)
     erv = eachrow(sdf)
@@ -119,7 +119,7 @@ end
 end
 
 @testset "getproperty and propertynames" begin
-    df_base = DataFrame([11:16 21:26 31:36 41:46])
+    df_base = DataFrame(Tables.table([11:16 21:26 31:36 41:46]))
     for df in (df_base, view(df_base, 1:3, 1:3))
         for x in (eachcol(df), eachrow(df))
             @test propertynames(x) == propertynames(df)
@@ -138,7 +138,7 @@ end
 end
 
 @testset "keys, values and pairs for eachcol" begin
-    df = DataFrame([11:16 21:26 31:36 41:46])
+    df = DataFrame(Tables.table([11:16 21:26 31:36 41:46]))
 
     cols = eachcol(df)