+++ test/reshape.jl
@@ -84,7 +84,7 @@ const ≅ = isequal
     @test_throws ArgumentError unstack(df, Int[], :Key, :Value)
     @test_throws ArgumentError unstack(df, r"xxxxx", :Key, :Value)
     @test_throws ArgumentError unstack(df, Symbol[], :Key, :Value)
-    @test_throws ArgumentError unstack(stack(DataFrame(rand(10, 10))),
+    @test_throws ArgumentError unstack(stack(DataFrame(Tables.table(rand(10, 10)))),
                                   :id, :variable, :value)
     @test_throws TypeError unstack(df, :Key, :Value, renamecols=Symbol)
 
@@ -187,7 +187,7 @@ end
 
 @testset "stack-unstack correctness" begin
     Random.seed!(1234)
-    x = DataFrame(rand(100, 50))
+    x = DataFrame(Tables.table(rand(100, 50), header=[Symbol(:x, i) for i in 1:50]))
     x[!, :id] = [1:99; missing]
     x[!, :id2] = string.("a", x[!, :id])
     x[!, :s] = [i % 2 == 0 ? randstring() : missing for i in 1:100]
@@ -329,13 +329,13 @@ end
     @test d1us3 == unstack(d1s2)
 
     # test unstack with exactly one key column that is not passed
-    df1 = stack(DataFrame(rand(10,10)))
+    df1 = stack(DataFrame(Tables.table(rand(10,10))))
     df1[!, :id] = 1:100
     @test size(unstack(df1, :variable, :value)) == (100, 11)
     @test unstack(df1, :variable, :value) ≅ unstack(df1)
 
     # test empty keycol
-    @test_throws ArgumentError unstack(stack(DataFrame(rand(3,2))), :variable, :value)
+    @test_throws ArgumentError unstack(stack(DataFrame(Tables.table(rand(3,2)))), :variable, :value)
 end
 
 @testset "column names duplicates" begin
@@ -494,7 +494,7 @@ end
 end
 
 @testset "test stack eltype" begin
-    df = DataFrame(rand(4,5))
+    df = DataFrame(Tables.table(rand(4,5)))
     sdf = stack(df)
     @test eltype(sdf.variable) === String
     @test eltype(typeof(sdf.variable)) === String